Made mapping part of the class.
Can either pass a higher-order mapping with the constructor, or use the default which is MappingQ1 (i.e. as it was originally).

In order to do this, I overloaded the constructor and created a setup_constructor function which performs the previous functions which were inside the main body of the constructor function.